Add function for sorting array with even and odd numbers
Implemented a function that takes two parameters, the first being an array of integers, and the second being the number of elements in the array. The function creates a vector of integers containing the same elements as the given array, but with even numbers preceding odd numbers (the relative order of even and odd numbers should be the same as in the original array). For example, if the array contains elements 3, 5, 2, 7, 6, 4, 1, 8, 5, 9, 4, 3, and 4, the returned vector should contain elements 2, 6, 4, 8, 4, 4, 3, 5, 7, 1, 5, 9, and 3. The main program demonstrates the function by first taking input from the keyboard for an array of capacity 10 elements. After that, the program calls the implemented function to generate a vector with elements ordered as described above relative to the elements of the entered array. Finally, the program prints the elements of the generated vector on the screen, separated by spaces.