Function for counting digits in vector elements
Function that takes a vector of integers as a parameter and returns a new vector whose elements represent the number of digits in the corresponding elements of the input vector. For example, if the function is passed a vector with elements 32, 459, 72, 8, 24, 771, and 13, the function should return a vector with elements 2, 3, 2, 1, 2, 3, and 2. The main program demonstrates the function with a short test program where a natural number 𝑛 is first entered from the keyboard, followed by 𝑛 vector elements.