Implemented function for digit sum vector creation
Function that takes a vector of integers as a parameter and returns a new vector whose elements are the sums of the digits of the corresponding elements in the input vector. For instance, if the function is given a vector with elements 32, 459, 72, 8, 24, 771, and 13, it will return a vector with elements 5, 18, 9, 8, 6, 15, and 4.