Function for sorting matrix by product of their rows
Generic function SortirajPoProizvoduRedova, which takes a matrix organized as a vector of vectors of arbitrary elements, capable of multiplication and comparison. The function transforms the matrix to be sorted by the product of rows in ascending order. The function utilizes the sort function from the algorithm library, passing a generic criterion function named Kriterij.