Function to check prime numbers in a vector without calling any external function
A function named 'check_prime_vector' that takes a vector of integers as input and returns a vector of boolean values of the same size. The function sets the i-th element of the result vector to 'true' if the i-th element of the input vector is a prime number, otherwise sets it to 'false'. No external helper functions are used within this implementation.